manage component
