+++ src/app/app.component.css
@@ -1,0 +1,4 @@
+:host {
+  display: flex;
+  flex: 1;
+}
