+++ src/app/manage/manage.component.css
@@ -1,0 +1,18 @@
+.mat-form-field {
+  font-size: 14px;
+  width: 100%;
+}
+
+.content {
+  padding: 12px;
+}
+
+.fab-bottom-right {
+  position: fixed;
+  right: 16px;
+  bottom: 16px;
+}
+
+/deep/ .mat-list-item-content {
+  height: auto !important;
+}
