+++ src/app/menu/menu.component.css
@@ -1,51 +1,13 @@
-.example-container {
-  display: flex;
-  flex-direction: column;
-  min-width: 300px;
-}
-
-.example-header {
-  min-height: 64px;
-  padding: 8px 24px 0;
-}
-
-.mat-form-field {
-  font-size: 14px;
-  width: 100%;
-}
-
-.mat-table {
-  overflow: auto;
-  max-height: 500px;
-}
-
 :host {
   display: flex;
   flex: 1;
 }
 
-mat-sidenav {
-  width: 320px;
-}
+/*mat-sidenav {*/
+  /*width: 320px;*/
+/*}*/
 
 .content {
   padding: 12px;
 }
 
-/deep/ mat-icon.avatar {
-  overflow: hidden;
-  width: 64px;
-  height: 64px;
-  border-radius: 50%;
-  margin: 12px;
-}
-
-.fab-bottom-right {
-  position: fixed;
-  right: 16px;
-  bottom: 16px;
-}
-
-/deep/ .mat-list-item-content {
-  height: auto !important;
-}
